+++ src/main/java/io/supertokens/ActiveUsers.java
@@ -1,5 +1,6 @@
 package io.supertokens;
 
+import io.supertokens.pluginInterface.ActiveUsersStorage;
 import io.supertokens.pluginInterface.Storage;
 import io.supertokens.pluginInterface.StorageUtils;
 import io.supertokens.pluginInterface.authRecipe.sqlStorage.AuthRecipeSQLStorage;
@@ -7,6 +8,7 @@ import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
 import io.supertokens.pluginInterface.multitenancy.AppIdentifier;
 import io.supertokens.pluginInterface.multitenancy.exceptions.TenantOrAppNotFoundException;
+import io.supertokens.pluginInterface.sqlStorage.SQLStorage;
 import io.supertokens.storageLayer.StorageLayer;
 import org.jetbrains.annotations.TestOnly;
 
@@ -35,6 +37,20 @@ public class ActiveUsers {
             throws StorageQueryException, TenantOrAppNotFoundException {
         Storage storage = StorageLayer.getStorage(appIdentifier.getAsPublicTenantIdentifier(), main);
         return StorageUtils.getActiveUsersStorage(storage).countUsersActiveSince(appIdentifier, time);
+    }
+
+    public static void updateLastActiveAfterLinking(Main main, AppIdentifier appIdentifier, String primaryUserId,
+                                                    String recipeUserId)
+            throws StorageQueryException, TenantOrAppNotFoundException, StorageTransactionLogicException {
+        ActiveUsersStorage activeUsersStorage =
+                StorageUtils.getActiveUsersStorage(StorageLayer.getStorage(appIdentifier.getAsPublicTenantIdentifier(), main));
+
+        ((SQLStorage) activeUsersStorage).startTransaction(con -> {
+            activeUsersStorage.deleteUserActive_Transaction(con, appIdentifier, recipeUserId);
+            return null;
+        });
+
+        updateLastActive(appIdentifier, main, primaryUserId);
     }
 
     @TestOnly